requirements.txt
asgiref==3.3.4
attrs==21.2.0
cffi==1.14.5
cryptography==3.4.7
Django==3.2.4
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
drf-spectacular==0.17.2
gunicorn==20.0.4
inflection==0.5.1
jsonschema==3.2.0
Pillow==8.2.0
pycparser==2.20
PyMySQL==1.0.2
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
